+++ pytest_cache_assert/_check_assert/serializer.py
@@ -1,15 +1,18 @@
 """Provide additional encoding functionality to check_assert."""
 
 import re
+from datetime import datetime
 from enum import Enum
 from json import JSONEncoder
 from pathlib import Path, PurePath
+from uuid import UUID
 
 import punq
 from beartype import beartype
-from beartype.typing import Any, Callable, Dict, Pattern, Union
+from beartype.typing import Any, Callable, Dict, List, Pattern, Union, Iterable
 
 from .config import CacheAssertContainerKeys, cache_assert_container
+from .constants import DIFF_TYPES
 
 _RE_MEMORY_ADDRESS = re.compile(r' at 0x[^>]+>')
 """Regex for matching the hex memory address in a function signature."""
@@ -41,7 +44,7 @@ def coerce_if_known_type(value: Any) -> Any:
         ((Path, PurePath), Path.as_posix),
         (Callable, replace_memory_address),
         (Enum, lambda _e: _e.value),
-        (Pattern, str),
+        ((datetime, UUID, Pattern), str),
         (complex, lambda _o: [_o.real, _o.imag]),
     ])
     for _types, func in rules:
@@ -52,20 +55,38 @@ def coerce_if_known_type(value: Any) -> Any:
 
 
 @beartype
-def recursive_serialize(value: Any) -> Union[Dict[str, Any], str]:
+def recursive_data_converter(data: Any) -> DIFF_TYPES:
+    """Recursively Coerce new data to type that can be compared.
+
+    Args:
+        data: data to coerce
+
+    Returns:
+        DIFF_TYPES: DiffResult-safe data
+
+    """
+    data = coerce_if_known_type(data)
+
+    if isinstance(data, (tuple, List)):
+        return [*map(coerce_if_known_type, data)]  # For performance, only provides one level of recursion
+    if isinstance(data, dict) and data:
+        return {_k: recursive_data_converter(_v) for _k, _v in data.items()}
+
+    return data
+
+
+@beartype
+def recursive_serialize(value: Any) -> Any:
     """Recursive serialization function.
 
     Args:
         value: value to serialize
 
     Returns:
-        Union[Dict[str, Any], str]: JSON-safe data
+        Union[Dict[str, Any], str]: JSON-safe data (numbers, etc.)
 
     """
-    if isinstance(value, dict) and value:
-        return {_k: recursive_serialize(_v) for _k, _v in value.items()}
-
-    return str(coerce_if_known_type(value))
+    return recursive_data_converter(value)
 
 
 class CacheAssertSerializer(JSONEncoder):