+++ docs/docs/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Fix
+
+- handle one-level of recursion in lists
+
 ## 1.3.1 (2022-02-27)
 
 ### Fix